split CskgCsvLoader and delegate
